bloomfilter: combine a couple modules into one
merge Data.BloomFilter.Mutable.Internal into Data.BloomFilter.Mutable
